app/requirements.txt

streamlit
torch
torchvision
timm
peft
Pillow
numpy
